requirements.txt
# Discord Bot Dependencies
discord.py>=2.4.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0

# Development Dependencies (optional)
black>=23.0.0
flake8>=6.0.0
